Code refinement - updated fromMap's completed_at to get dataTime value in local time zone. - created a class for getting completed task by userId for the chart_screen. - fixed code where initially when completed code number is increased in last week instead of this week. - changed where status should change to In Progress when redirecting from Pomodoro Timer to Edit Page if task is not completed, it does not do that initially. - made home page to refresh every time user comes back from other page, to check for any updates made. Big Changes: - renamed project's app to new title TomatoTask - gave app an actual icon

+++ lib/data/repo/task_repo_supabase.dart
@@ -1,5 +1,5 @@
 import 'package:supabase_flutter/supabase_flutter.dart';
-import 'package:taskapp_with_pomodoro/data/model/task.dart';
+import 'package:tomato_task/data/model/task.dart';
 
 class TaskRepoSupabase {
   static final TaskRepoSupabase _instance = TaskRepoSupabase._init();
@@ -31,23 +31,25 @@ class TaskRepoSupabase {
   Future<void> updateTask(Task task) async {
     final updatedMap = task.toMap();
 
-    if (task.completedAt != null) {
-      updatedMap['completed_at'] = task.completedAt!.millisecondsSinceEpoch;
-    } else {
-      updatedMap['completed_at'] = null;
-    }
+    updatedMap['completed_at'] = task.completedAt?.toIso8601String();
     await supabase
-    .from('tasks')
-    .update(task.toMap())
-    .eq('id', task.id!)
-    .eq('user_id', task.userId!);
+        .from('tasks')
+        .update(task.toMap())
+        .eq('id', task.id!)
+        .eq('user_id', task.userId!);
   }
 
   Future<void> deleteTask(int id, String userId) async {
-    await supabase
-    .from('tasks')
-    .delete()
-    .eq('id', id)
-    .eq('user_id', userId);
+    await supabase.from('tasks').delete().eq('id', id).eq('user_id', userId);
+  }
+
+  Future<List<Task>> getCompletedTasksByUser(String userId) async {
+    final response = await supabase
+        .from('tasks')
+        .select()
+        .eq('user_id', userId)
+        .eq('status', 'completed')
+        .order('completed_at', ascending: false);
+    return response.map((json) => Task.fromMap(json)).toList();
   }
 }
